+++ Sources/Sentry/SentryDataCategoryMapper.m
@@ -11,6 +11,7 @@ NSString *const kSentryDataCategoryNameTransaction = @"transaction";
 NSString *const kSentryDataCategoryNameAttachment = @"attachment";
 NSString *const kSentryDataCategoryNameUserFeedback = @"user_report";
 NSString *const kSentryDataCategoryNameProfile = @"profile";
+NSString *const kSentryDataCategoryNameReplay = @"replay";
 NSString *const kSentryDataCategoryNameStatsd = @"statsd";
 NSString *const kSentryDataCategoryNameUnknown = @"unknown";
 
@@ -33,6 +34,9 @@ sentryDataCategoryForEnvelopItemType(NSString *itemType)
     }
     if ([itemType isEqualToString:SentryEnvelopeItemTypeProfile]) {
         return kSentryDataCategoryProfile;
+    }
+    if ([itemType isEqualToString:SentryEnvelopeItemTypeReplayVideo]) {
+        return kSentryDataCategoryReplay;
     }
     if ([itemType isEqualToString:SentryEnvelopeItemTypeStatsd]) {
         return kSentryDataCategoryStatsd;
@@ -77,6 +81,9 @@ sentryDataCategoryForString(NSString *value)
     if ([value isEqualToString:kSentryDataCategoryNameProfile]) {
         return kSentryDataCategoryProfile;
     }
+    if ([value isEqualToString:kSentryDataCategoryNameReplay]) {
+        return kSentryDataCategoryReplay;
+    }
     if ([value isEqualToString:kSentryDataCategoryNameStatsd]) {
         return kSentryDataCategoryStatsd;
     }
@@ -112,6 +119,8 @@ nameForSentryDataCategory(SentryDataCategory category)
         return kSentryDataCategoryNameStatsd;
     case kSentryDataCategoryUnknown:
         return kSentryDataCategoryNameUnknown;
+    case kSentryDataCategoryReplay:
+        return kSentryDataCategoryNameReplay;
     }
 }