+++ Sources/Sentry/include/SentryDataCategory.h
@@ -15,7 +15,8 @@ typedef NS_ENUM(NSUInteger, SentryDataCategory) {
     kSentryDataCategoryUserFeedback = 6,
     kSentryDataCategoryProfile = 7,
     kSentryDataCategoryStatsd = 8,
-    kSentryDataCategoryUnknown = 9
+    kSentryDataCategoryReplay = 9,
+    kSentryDataCategoryUnknown = 10
 };
 
 static DEPRECATED_MSG_ATTRIBUTE(
@@ -31,5 +32,6 @@ static DEPRECATED_MSG_ATTRIBUTE(
           @"user_report",
           @"profile",
           @"statsd",
+          @"replay",
           @"unkown",
       };