+++ Tests/SentryTests/SentryOptionsTest.m
@@ -544,6 +544,7 @@
 #if SENTRY_HAS_UIKIT
         @"enableUIViewControllerTracing" : [NSNull null],
         @"attachScreenshot" : [NSNull null],
+        @"sessionReplayOptions" : [NSNull null],
 #endif
         @"enableAppHangTracking" : [NSNull null],
         @"appHangTimeoutInterval" : [NSNull null],
@@ -603,6 +604,9 @@
     XCTAssertEqual(options.enableUserInteractionTracing, YES);
     XCTAssertEqual(options.enablePreWarmedAppStartTracing, NO);
     XCTAssertEqual(options.attachViewHierarchy, NO);
+    if (@available(iOS 16.0, tvOS 16.0, *)) {
+        XCTAssertNil(options.sessionReplayOptions);
+    }
 #endif
     XCTAssertFalse(options.enableTracing);
     XCTAssertTrue(options.enableAppHangTracking);
@@ -776,6 +780,27 @@
 - (void)testEnablePreWarmedAppStartTracking
 {
     [self testBooleanField:@"enablePreWarmedAppStartTracing" defaultValue:NO];
+}
+
+- (void)testSessionReplaySettingsInit
+{
+    if (@available(iOS 16.0, tvOS 16.0, *)) {
+        SentryOptions *options = [self getValidOptions:@{
+            @"sessionReplayOptions" :
+                @ { @"replaysSessionSampleRate" : @2, @"replaysOnErrorSampleRate" : @4 }
+        }];
+        XCTAssertEqual(options.sessionReplayOptions.replaysSessionSampleRate, 2);
+        XCTAssertEqual(options.sessionReplayOptions.replaysOnErrorSampleRate, 4);
+    }
+}
+
+- (void)testSessionReplaySettingsDefaults
+{
+    if (@available(iOS 16.0, tvOS 16.0, *)) {
+        SentryOptions *options = [self getValidOptions:@{ @"sessionReplayOptions" : @ {} }];
+        XCTAssertEqual(options.sessionReplayOptions.replaysSessionSampleRate, 0);
+        XCTAssertEqual(options.sessionReplayOptions.replaysOnErrorSampleRate, 0);
+    }
 }
 
 #endif