+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Features
 
+- Add Session Replay (#3625)
 - Add support for Sentry [Spotlight](https://spotlightjs.com/) (#3642), which is basically Sentry
 for development. Read our [blog post](https://blog.sentry.io/sentry-for-development/) to find out more.
 - Add field `SentrySDK.detectedStartUpCrash` (#3644)